Refatoracao: Eliminando necessidade de rotas mutiplas para vagas
