+++ backend/src/app/models/Vacancy.js
@@ -5,23 +5,9 @@ Base.init({ table: 'vacancies' })
 
 module.exports = {
     ...Base,
-    // async search({ filter }) {
+    async search({ filter }) {
 
-    //     let query = ` SELECT * FROM vacancies order by` + filter + `desc;`
-
-    //     const results = await db.query(query)
-    //     return results.rows
-    // }
-    async searchQuantity({ }) {
-
-        let query = ` SELECT * FROM vacancies order by quantity desc;`
-
-        const results = await db.query(query)
-        return results.rows
-    },
-    async searchSalary({ }) {
-
-        let query = ` SELECT * FROM vacancies order by salary desc;`
+        let query = ` SELECT * FROM vacancies order by ${filter} desc;`
 
         const results = await db.query(query)
         return results.rows
