--- backend/src/routes/vacancyRoute.js
@@ -1,9 +1,0 @@
-const express = require('express')
-const routes = express.Router()
-
-const VacancyController = require('../app/controllers/VacancyController')
-
-routes.get("/", VacancyController.indexQuantity)
-routes.get("/", VacancyController.indexSalary)
-
-module.exports = routes
