--- backend/src/routes/vacancyRouteQuantity.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const routes = express.Router()
-
-const VacancyController = require('../app/controllers/VacancyController')
-
-routes.get("/", VacancyController.indexQuantity)
-
-module.exports = routes
