+++ backend/src/server.js
@@ -3,7 +3,7 @@ const express = require('express')  // importamos o modulo express (servidor) na
 const routes = require('./routes')
 
 const server = express() // criação do servidor na variável server
-const port = process.env.PORT || 3000
+const port = process.env.PORT || 3001
 
 server.use(express.json()) //recognize the incoming Request Object as a JSON Object
 server.use(express.urlencoded({ extended: true })) //recognize the incoming Request Object as strings or arrays. urlencoded é um parser das informações vindas no corpo da requisição.
